Refactor Object sstableCompressionRatio to double

+++ src/java/org/apache/cassandra/tools/nodetool/stats/StatsTableComparator.java
@@ -326,8 +326,8 @@ public class StatsTableComparator implements Comparator<StatsTable>
         }
         else if (sortKey.equals("sstable_compression_ratio"))
         {
-            result = compareDoubles((Double) stx.sstableCompressionRatio,
-                                    (Double) sty.sstableCompressionRatio);
+            result = compareDoubles(stx.sstableCompressionRatio,
+                                    sty.sstableCompressionRatio);
         }
         else if (sortKey.equals("sstable_count"))
         {
